fix(security): inline redact into appendSessionTranscriptMessage
All 4 callsites of appendSessionTranscriptMessage (transcript.ts,
attempt-execution.ts, chat-transcript-inject.ts, transcript-mirror.ts)
now redact sensitive content before writing to disk, closing the path-B
leak identified in #64046.

Implementation:
- Extract redact helpers from session-tool-result-guard-wrapper.ts into
  new src/agents/transcript-redact.ts (zero logic change, shared module)
- Widen appendSessionTranscriptMessage config param from
  SessionWriteLockAcquireTimeoutConfig to OpenClawConfig (structural
  superset, no callsite changes required)
- Call redactTranscriptMessage before building the JSONL entry; falls
  back to resolveConfigRedaction() when config is undefined

Fixes #64046
