+++ src/config/sessions/transcript-append.ts
@@ -2,11 +2,13 @@ import { randomUUID } from "node:crypto";
 import fs from "node:fs/promises";
 import path from "node:path";
 import { StringDecoder } from "node:string_decoder";
+import type { AgentMessage } from "@mariozechner/pi-agent-core";
 import {
   acquireSessionWriteLock,
-  type SessionWriteLockAcquireTimeoutConfig,
   resolveSessionWriteLockAcquireTimeoutMs,
 } from "../../agents/session-write-lock.js";
+import { redactTranscriptMessage } from "../../agents/transcript-redact.js";
+import type { OpenClawConfig } from "../../config/types.openclaw.js";
 
 const TRANSCRIPT_APPEND_SCAN_CHUNK_BYTES = 64 * 1024;
 const SESSION_MANAGER_APPEND_MAX_BYTES = 8 * 1024 * 1024;
@@ -236,7 +238,7 @@ export async function appendSessionTranscriptMessage(params: {
   sessionId?: string;
   cwd?: string;
   useRawWhenLinear?: boolean;
-  config?: SessionWriteLockAcquireTimeoutConfig;
+  config?: OpenClawConfig;
 }): Promise<{ messageId: string }> {
   return await withTranscriptAppendQueue(params.transcriptPath, () =>
     appendSessionTranscriptMessageLocked(params),
@@ -250,7 +252,7 @@ async function appendSessionTranscriptMessageLocked(params: {
   sessionId?: string;
   cwd?: string;
   useRawWhenLinear?: boolean;
-  config?: SessionWriteLockAcquireTimeoutConfig;
+  config?: OpenClawConfig;
 }): Promise<{ messageId: string }> {
   const lock = await acquireSessionWriteLock({
     sessionFile: params.transcriptPath,
@@ -285,12 +287,13 @@ async function appendSessionTranscriptMessageLocked(params: {
         nonSessionEntryCount: leafInfo.nonSessionEntryCount,
       };
     }
+    const finalMessage = redactTranscriptMessage(params.message as AgentMessage, params.config);
     const entry = {
       type: "message",
       id: messageId,
       ...(shouldRawAppend ? {} : { parentId: leafInfo.leafId ?? null }),
       timestamp: new Date(now).toISOString(),
-      message: params.message,
+      message: finalMessage,
     };
     await fs.appendFile(params.transcriptPath, `${JSON.stringify(entry)}\n`, "utf-8");
     return { messageId };
